Add files via upload
This project is a Student Information Form with MongoDB as the database to store submitted data. It features a simple HTML form where students enter their Registration Number, Name, Email, and Branch. Upon submission, the data is sent to the server via a POST request, and the server processes and saves this information in a MongoDB database. MongoDB is chosen for its flexibility in handling unstructured data, allowing efficient storage and retrieval. The project also includes basic validation to ensure data accuracy and security, making it ideal for applications requiring student data management.